# docker/docs/swagger.yaml
definitions:
  users-service_app_models_ent.ActivationCode:
    properties:
      code:
        description: Code holds the value of the "code" field.
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.ActivationCodeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ActivationCodeQuery when eager-loading is set.
      expires_at:
        description: ExpiresAt holds the value of the "expires_at" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      used:
        description: Used holds the value of the "used" field.
        type: boolean
    type: object
  users-service_app_models_ent.ActivationCodeEdges:
    properties:
      owner:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.User'
        description: Owner holds the value of the owner edge.
    type: object
  users-service_app_models_ent.ResetCode:
    properties:
      code:
        description: Code holds the value of the "code" field.
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.ResetCodeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ResetCodeQuery when eager-loading is set.
      expires_at:
        description: ExpiresAt holds the value of the "expires_at" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
...
  users-service_app_models_ent.SecondFactorCode:
    properties:
      code:
        description: Code holds the value of the "code" field.
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.SecondFactorCodeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the SecondFactorCodeQuery when eager-loading is set.
      expires_at:
        description: ExpiresAt holds the value of the "expires_at" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      used:
        description: Used holds the value of the "used" field.
        type: boolean
    type: object
...
  users-service_app_models_ent.User:
    properties:
      active:
        description: Active holds the value of the "active" field.
        type: boolean
      blacklisted:
        description: Blacklisted holds the value of the "blacklisted" field.
        type: boolean
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      mail:
        description: Mail holds the value of the "mail" field.
        type: string
      password:
        description: Password holds the value of the "password" field.
        type: string
      phone:
        description: Phone holds the value of the "phone" field.
        type: string
      removed:
        description: Removed holds the value of the "removed" field.
        type: boolean
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      verified:
        description: Verified holds the value of the "verified" field.
        type: boolean
    type: object
  users-service_app_models_ent.UserAction:
    properties:
      action:
        description: Action holds the value of the "action" field.
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      details:
        description: Details holds the value of the "details" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserActionEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserActionQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      type:
        description: Type holds the value of the "type" field.
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
...
  users-service_app_models_ent.UserEdges:
    properties:
      activation_code:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.ActivationCode'
        description: ActivationCode holds the value of the activation_code edge.
      reset_code:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.ResetCode'
        description: ResetCode holds the value of the reset_code edge.
      second_factor_code:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.SecondFactorCode'
        description: SecondFactorCode holds the value of the second_factor_code edge.
      user_actions:
        description: UserActions holds the value of the user_actions edge.
        items:
          $ref: '#/definitions/users-service_app_models_ent.UserAction'
        type: array
      user_details:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserDetails'
        description: UserDetails holds the value of the user_details edge.
      user_devices:
        description: UserDevices holds the value of the user_devices edge.
        items:
          $ref: '#/definitions/users-service_app_models_ent.UserDevice'
        type: array
      user_roles:
        description: UserRoles holds the value of the user_roles edge.
        items:
          $ref: '#/definitions/users-service_app_models_ent.UserRole'
        type: array
      user_settings:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserSettings'
        description: UserSettings holds the value of the user_settings edge.
      verification_code:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.VerificationCode'
        description: VerificationCode holds the value of the verification_code edge.
    type: object
...
  users-service_app_models_requests.Code:
    properties:
      browser_name:
        type: string
      browser_version:
        type: string
      code:
        type: string
      device_token:
        type: string
      ip_address:
        type: string
      os_name:
        type: string
      os_version:
        type: string
      user_agent:
        type: string
    required:
    - device_token
    type: object
  users-service_app_models_requests.ConfirmPasswordReset:
    properties:
      browser_name:
        type: string
      browser_version:
        type: string
      code:
        type: string
      device_token:
        type: string
      ip_address:
        type: string
      os_name:
        type: string
      os_version:
        type: string
      password:
        type: string
      user_agent:
        type: string
    required:
    - device_token
    type: object
...
  users-service_app_models_requests.Login:
    properties:
      browser_name:
        type: string
      browser_version:
        type: string
      device_token:
        type: string
      ip_address:
        type: string
      mail:
        type: string
      os_name:
        type: string
      os_version:
        type: string
      password:
        type: string
      user_agent:
        type: string
    required:
    - device_token
    type: object
  users-service_app_models_requests.Mail:
    properties:
      browser_name:
        type: string
      browser_version:
        type: string
      device_token:
        type: string
      ip_address:
        type: string
      mail:
        type: string
      os_name:
        type: string
      os_version:
        type: string
      user_agent:
        type: string
    required:
    - device_token
    type: object
  users-service_app_models_requests.Register:
    properties:
      browser_name:
        type: string
      browser_version:
        type: string
      device_token:
        type: string
      ip_address:
        type: string
      mail:
        type: string
      name:
        type: string
      os_name:
        type: string
      os_version:
        type: string
      password:
        minLength: 8
        type: string
      user_agent:
        type: string
    required:
    - device_token
    - mail
    - password
    type: object
...
  users-service_app_models_responses.ErrorResponse:
    properties:
      error:
        type: string
      requestID:
        type: string
      status:
        type: string
    type: object
  users-service_app_models_responses.SuccessResponse:
    properties:
      data: {}
      requestID:
        type: string
      status:
        type: string
    type: object
  users-service_app_models_responses.User:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/users-service_app_models_ent.User'
    type: object
info:
  contact: {}
paths:
  /api/v1/auth/activation/account:
    patch:
      consumes:
      - application/json
      description: Activates a user's account using a provided activation code.
      parameters:
      - description: Activation code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users-service_app_models_requests.Code'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Account activated successfully
          schema:
            allOf:
            - $ref: '#/definitions/users-service_app_models_responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/users-service_app_models_responses.User'
              type: object
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "422":
          description: Unprocessable Entity - Invalid or expired code
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
      summary: Activate user account
      tags:
      - auth
  /api/v1/auth/activation/resend/code:
    post:
      consumes:
      - application/json
      description: Requests a new activation code to be sent to the user's email address.
      parameters:
      - description: User's email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users-service_app_models_requests.Mail'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Successfully requested a new activation code
          schema:
            allOf:
            - $ref: '#/definitions/users-service_app_models_responses.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "422":
          description: Unprocessable Entity - Service failed to process the request
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
      summary: Resend activation code
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Initiates a WebSocket connection to handle user login with real-time
        progress updates.
      parameters:
      - description: User login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users-service_app_models_requests.Login'
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
      summary: Login a user
      tags:
      - auth
...
  /api/v1/auth/password/reset/request:
    post:
      consumes:
      - application/json
      description: Initiates the password reset process by sending a code to the user's
        email.
      parameters:
      - description: User's email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users-service_app_models_requests.Mail'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Password reset email sent
          schema:
            allOf:
            - $ref: '#/definitions/users-service_app_models_responses.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "422":
          description: Unprocessable Entity - Service failed to process the request
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
      summary: Request password reset
      tags:
      - auth
...
  /api/v1/auth/password/reset/request/confirm:
    patch:
      consumes:
      - application/json
      description: Sets a new password for the user using a valid reset code.
      parameters:
      - description: Password reset code and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users-service_app_models_requests.ConfirmPasswordReset'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Password has been reset successfully
          schema:
            allOf:
            - $ref: '#/definitions/users-service_app_models_responses.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "422":
          description: Unprocessable Entity - Invalid code or other error
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
      summary: Confirm password reset
      tags:
      - auth
  /api/v1/auth/password/reset/request/resend:
    post:
      consumes:
      - application/json
      description: Requests a new password reset code to be sent to the user's email
        address.
      parameters:
      - description: User's email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users-service_app_models_requests.Mail'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Successfully requested a new reset code
          schema:
            allOf:
            - $ref: '#/definitions/users-service_app_models_responses.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "422":
          description: Unprocessable Entity - Service failed to process the request
            (e.g., user not found)
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
      summary: Resend Password Reset Code
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Initiates a WebSocket connection to handle user registration with
        real-time progress updates.
      parameters:
      - description: User registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users-service_app_models_requests.Register'
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
  /api/v1/auth/second-factor/resend/code:
    post:
      consumes:
      - application/json
      description: Requests a new second factor code to be sent to the user.
      parameters:
      - description: User's email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users-service_app_models_requests.Mail'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Successfully requested a new 2FA code
          schema:
            allOf:
            - $ref: '#/definitions/users-service_app_models_responses.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "422":
          description: Unprocessable Entity - Service failed to process the request
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
      summary: Resend second factor code
      tags:
      - auth
  /api/v1/auth/verification/account:
    patch:
      consumes:
      - application/json
      description: Verifies a user's account with a provided code, returning the user
        and a JWT.
      parameters:
      - description: Verification code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users-service_app_models_requests.Code'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Account verified successfully
          schema:
            allOf:
            - $ref: '#/definitions/users-service_app_models_responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/users-service_app_models_responses.User'
              type: object
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "422":
          description: Unprocessable Entity - Invalid or expired code
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
      summary: Verify user account (e.g., phone)
      tags:
      - auth
  /api/v1/auth/verification/resend/code:
    post:
      consumes:
      - application/json
      description: Requests a new verification code to be sent to the user.
      parameters:
      - description: User's email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users-service_app_models_requests.Mail'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Successfully requested a new verification code
          schema:
            allOf:
            - $ref: '#/definitions/users-service_app_models_responses.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "422":
          description: Unprocessable Entity - Service failed to process the request
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
      summary: Resend verification code
      tags:
      - auth
...
swagger: "2.0"